requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pillow>=10.0.0
requests>=2.28.0
google-genai>=1.0.0
